Deleted Choice, typedefed bool to cooperation

+++ main.cpp
@@ -1,6 +1,6 @@
 #include <iostream>
 
-#include "choice.h"
+//#include "choice.h"
 #include "player.h"
 #include "gameManager.h"
 
@@ -15,12 +15,11 @@ int main() {
     players.push_back(new LoyalPrisoner());
     players.push_back(new SelfishPrisoner());
 
-    const int n = players.size();
-    for (int i = 0; i < n; i++) { players[i]->print(); }
+    const size_t n = players.size();
+    for (size_t i = 0; i < n; i++) { players[i]->print(); }
 
     std::cout << std::endl;
 
-    //Player* player = players.at(0);
     vector<Player*>::iterator player;
     for (player = players.begin(); player != players.end(); player++) {
         (*player)->print();
